pyproject.toml
[project]
name = "ai-career-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "groq>=1.0.0",
    "openai>=2.21.0",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
